requirements.txt
asgiref==3.6.0
Django==4.1.4
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
greenlet==2.0.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
